Remove outdated version file and improve data processing
This commit makes three main changes:
1) The config/version.yml file is removed as it was outdated and not used elsewhere in the code.
2) Filament QuickCountResource file has been updated to group by administrative units. ManageHitungSuaraPartais has been renamed to ManageHitungSuaraPartai and updated to calculate percentage of votes automatically as soon as the number of party votes is entered. Dashboard widget has been updated to get the total number of candidates and votes, remaining votes and percentages.
3) Composer packages are updated to latest versions.

The changes were made to improve data processing efficiency in the dashboard and keep the code and dependencies tidy and up-to-date.

+++ app/Filament/Widgets/QuickCountStatistik.php
@@ -31,7 +31,7 @@ class QuickCountStatistik extends BaseWidget
                 ->description('Total Jumlah Suara Calon')
                 ->descriptionIcon('heroicon-m-calculator')
                 ->color('warning'),
-            Stat::make('Suara Partai', Helpers::shortNumber(HitungSuaraPartai::sum('jumlah_suara_partai')))
+            Stat::make('Suara Partai', Helpers::shortNumber((float) HitungSuaraPartai::sum('jumlah_suara_partai')))
                 ->description('Total Suara Partai')
                 ->descriptionIcon('heroicon-m-calculator')
                 ->color('info'),
